Changed the behavior of kacanopen (effects the Servo Cylinder actuators).
If the actuators are given a new position setpoint to move to, they will continue to their previous setpoint, before moving to the new one. If both setpoints take the 402 device in the same direction, it will not pause at the previous position before moving to the new one. (i.e. Bit 9 is now set in the 402 controlword.)

Also, commented out the kacanopen console message that is printed for each device heartbeat.